pyproject.toml
[project]
name = "investor-agent"
version = "0.1.0"
description = "AI investment analysis agent that combines comprehensive financial data with professional investment research techniques"
authors = [{ name = "Paras Chamoli", email = "paras@getbindu.com" }]
readme = "README.md"
keywords = ['python', 'investment', 'finance', 'trading', 'agno', 'bindu', 'stock-market', 'financial-analysis', 'investment-research', 'mcp', 'model context protocol', 'technical analysis', 'yfinance']
requires-python = ">=3.12,<3.13"
dependencies = [
    "bindu==2026.9.4",
    "agno>=2.2.0",
    "openai>=2.11.0",
    "mem0ai>=1.0.1",
    "python-dotenv>=1.0.1",
    # Investment analysis dependencies (from investor-agent-org)
    "anysqlite>=0.0.5",
    "hishel[async]>=1.0.0",
    "numpy>=2.0",
    "html5lib>=1.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "pytrends>=4.9.2",
    "requests-cache>=1.2.1",
    "tenacity>=9.1.2",
    "yfinance[nospam]>=0.2.66",
]
classifiers = [
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

[project.urls]
Homepage = "https://Paraschamoli.github.io/investor-agent/"
Repository = "https://github.com/Paraschamoli/investor-agent"
Documentation = "https://Paraschamoli.github.io/investor-agent/"

[project.optional-dependencies]
ta = [
    "ta-lib>=0.6.8",
]

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["investor_agent"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "RUF", # ruff-specific
    "UP",  # pyupgrade
    "S",   # flake8-bandit
]
ignore = [
    "E501",  # Line too long
    "E731",  # Do not assign a lambda expression
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["investor_agent"]
